requirements.txt
pyrogram
tgcrypto
pykeyboard
python-dotenv
uvloop
telegraph
aiohttp
lxml
requests